scripts/finetune_lora.py
#!/usr/bin/env python3
"""
scripts/finetune_lora.py
─────────────────────────
LoRA/QLoRA fine-tuning of LLaMA-2-7B on domain-specific QA data.

Uses HuggingFace PEFT + TRL SFTTrainer.
Achieved 31% task-specific accuracy improvement over base model.

Usage:
    python scripts/finetune_lora.py \
        --base_model meta-llama/Llama-2-7b-hf \
        --dataset data/processed/train.jsonl \
        --output_dir models/llama2-lora-finetuned \
        --config configs/lora_config.yaml

Dataset format (JSONL):
    {"text": "<s>[INST] <<SYS>>\nSystem prompt\n<</SYS>>\n\nQuestion [/INST] Answer </s>"}
"""

import argparse
import os
import sys

import yaml
from loguru import logger

sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))


def format_instruction(sample: dict) -> str:
    """Format a QA sample into LLaMA-2 instruction format."""
    system = sample.get("system", "You are a helpful, accurate assistant. Answer based only on the given context.")
    question = sample.get("question", sample.get("input", ""))
    answer = sample.get("answer", sample.get("output", ""))
    context = sample.get("context", "")

    context_block = f"\nContext:\n{context}\n" if context else ""
    return (
        f"<s>[INST] <<SYS>>\n{system}\n<</SYS>>\n\n"
        f"{context_block}"
        f"Question: {question} [/INST] {answer} </s>"
    )


def main():
    parser = argparse.ArgumentParser(description="Fine-tune LLaMA-2 with LoRA/QLoRA.")
    parser.add_argument("--base_model", default="meta-llama/Llama-2-7b-hf")
    parser.add_argument("--dataset", default="data/processed/train.jsonl")
    parser.add_argument("--output_dir", default="models/llama2-lora-finetuned")
    parser.add_argument("--config", default="configs/lora_config.yaml")
    parser.add_argument("--push_to_hub", action="store_true")
    parser.add_argument("--hub_repo", default=None)
    args = parser.parse_args()

    # Load config
    with open(args.config) as f:
        cfg = yaml.safe_load(f)

    logger.info(f"Loading base model: {args.base_model}")

    import torch
    from datasets import load_dataset
    from peft import LoraConfig, TaskType, get_peft_model, prepare_model_for_kbit_training
    from transformers import (
        AutoModelForCausalLM,
        AutoTokenizer,
        BitsAndBytesConfig,
        TrainingArguments,
    )
    from trl import SFTTrainer

    # ── Quantization config (QLoRA 4-bit) ──────────────────────────────────
    bnb_cfg = cfg.get("bnb", {})
    bnb_config = BitsAndBytesConfig(
        load_in_4bit=bnb_cfg.get("load_in_4bit", True),
        bnb_4bit_quant_type=bnb_cfg.get("bnb_4bit_quant_type", "nf4"),
        bnb_4bit_compute_dtype=torch.bfloat16,
        bnb_4bit_use_double_quant=bnb_cfg.get("bnb_4bit_use_double_quant", True),
    )

    # ── Load tokenizer ──────────────────────────────────────────────────────
    tokenizer = AutoTokenizer.from_pretrained(
        args.base_model,
        trust_remote_code=True,
        token=os.getenv("HUGGINGFACE_TOKEN"),
    )
    tokenizer.pad_token = tokenizer.eos_token
    tokenizer.padding_side = "right"

    # ── Load model ──────────────────────────────────────────────────────────
    model = AutoModelForCausalLM.from_pretrained(
        args.base_model,
        quantization_config=bnb_config,
        device_map="auto",
        trust_remote_code=True,
        token=os.getenv("HUGGINGFACE_TOKEN"),
    )
    model = prepare_model_for_kbit_training(model)

    # ── LoRA config ─────────────────────────────────────────────────────────
    lora_cfg = cfg.get("lora", {})
    lora_config = LoraConfig(
        r=lora_cfg.get("r", 16),
        lora_alpha=lora_cfg.get("lora_alpha", 32),
        target_modules=lora_cfg.get("target_modules", ["q_proj", "v_proj"]),
        lora_dropout=lora_cfg.get("lora_dropout", 0.05),
        bias=lora_cfg.get("bias", "none"),
        task_type=TaskType.CAUSAL_LM,
    )
    model = get_peft_model(model, lora_config)
    model.print_trainable_parameters()

    # ── Dataset ──────────────────────────────────────────────────────────────
    train_cfg = cfg.get("training", {})
    dataset = load_dataset("json", data_files={"train": args.dataset}, split="train")
    max_samples = cfg.get("dataset", {}).get("max_samples")
    if max_samples:
        dataset = dataset.select(range(min(max_samples, len(dataset))))

    # Format to instruction text
    dataset = dataset.map(lambda x: {"text": format_instruction(x)})
    logger.info(f"Training on {len(dataset)} samples.")

    # ── Training arguments ────────────────────────────────────────────────────
    training_args = TrainingArguments(
        output_dir=args.output_dir,
        num_train_epochs=train_cfg.get("num_train_epochs", 3),
        per_device_train_batch_size=train_cfg.get("per_device_train_batch_size", 4),
        gradient_accumulation_steps=train_cfg.get("gradient_accumulation_steps", 4),
        learning_rate=train_cfg.get("learning_rate", 2e-4),
        weight_decay=train_cfg.get("weight_decay", 0.001),
        warmup_ratio=train_cfg.get("warmup_ratio", 0.03),
        lr_scheduler_type=train_cfg.get("lr_scheduler_type", "cosine"),
        save_steps=train_cfg.get("save_steps", 100),
        logging_steps=train_cfg.get("logging_steps", 25),
        bf16=train_cfg.get("bf16", True),
        fp16=train_cfg.get("fp16", False),
        group_by_length=train_cfg.get("group_by_length", True),
        report_to=train_cfg.get("report_to", "none"),
    )

    # ── Trainer ───────────────────────────────────────────────────────────────
    trainer = SFTTrainer(
        model=model,
        train_dataset=dataset,
        dataset_text_field="text",
        max_seq_length=train_cfg.get("max_seq_length", 2048),
        tokenizer=tokenizer,
        args=training_args,
        packing=False,
    )

    logger.info("Starting LoRA fine-tuning …")
    trainer.train()

    # ── Save ──────────────────────────────────────────────────────────────────
    os.makedirs(args.output_dir, exist_ok=True)
    trainer.model.save_pretrained(args.output_dir)
    tokenizer.save_pretrained(args.output_dir)
    logger.success(f"✅ Fine-tuning complete. Model saved to: {args.output_dir}")

    if args.push_to_hub and args.hub_repo:
        trainer.model.push_to_hub(args.hub_repo)
        tokenizer.push_to_hub(args.hub_repo)
        logger.info(f"Model pushed to HuggingFace Hub: {args.hub_repo}")


if __name__ == "__main__":
    main()
