requirements.txt
# Core LangChain
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langchain-core==0.2.38

# LLM & Embeddings
openai==1.40.0
tiktoken==0.7.0

# HuggingFace / Fine-tuning
transformers==4.44.0
peft==0.12.0
trl==0.9.6
bitsandbytes==0.43.3
accelerate==0.33.0
datasets==2.21.0
huggingface-hub==0.24.6
sentencepiece==0.2.0
protobuf==4.25.4

# Vector Store & Retrieval
faiss-cpu==1.8.0
rank-bm25==0.2.2

# Deep Learning
torch==2.4.0
torchvision==0.19.0

# API Framework
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9

# AWS S3
boto3==1.35.0
botocore==1.35.0

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.0
beautifulsoup4==4.12.3

# Evaluation
ragas==0.1.18
rouge-score==0.1.2
nltk==3.9.1

# Search Tools
duckduckgo-search==6.2.13
google-search-results==2.4.2  # SerpAPI

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
loguru==0.7.2
tenacity==8.5.0
httpx==0.27.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
httpx==0.27.0

# Guardrails
detoxify==0.5.2
